Fold the Select demo into the shared inputs template
