--- client-app/src/desktop/tabs/forms/SelectPanel.scss
@@ -1,10 +1,0 @@
-.tb-select-panel {
-  &__row {
-    margin-bottom: var(--xh-pad-double-px);
-    flex-shrink: 0;
-  }
-
-  &__label {
-    font-weight: 500;
-  }
-}
