+++ client-app/src/desktop/tabs/forms/index.ts
@@ -1,6 +1,5 @@
 export * from './FormPanel';
 export * from './DateRangePickerPanel';
 export * from './PickerPanel';
-export * from './SelectPanel';
 export * from './ToolbarFormPanel';
 export * from './inputs';
