+++ client-app/src/desktop/tabs/forms/inputs/index.ts
@@ -13,3 +13,4 @@ export * from './CodeInputsPanel';
 export * from './TogglesPanel';
 export * from './IntentInputPanel';
 export * from './SliderPanel';
+export * from './SelectPanel';
